+++ src/windowed_renderer.hpp
@@ -16,20 +16,23 @@
 #pragma once
 
 #include <videoviewer/videoviewer.hpp>
-#include <thread>
 
 #include "shared_resources.hpp"
 
-class RxRenderer
+#include <iostream>
+#include <thread>
+#include <atomic>
+
+class WindowedRenderer
 {
 public:
-    RxRenderer(std::string window_title, int window_width, int window_height, int framerate_ms);
-    ~RxRenderer();
+    WindowedRenderer(std::string window_title, int window_width, int window_height, int framerate_ms, std::atomic_bool& stop_is_requested);
+    ~WindowedRenderer();
     
-    RxRenderer(const RxRenderer&) = delete;
-    RxRenderer& operator=(const RxRenderer&) = delete;
-    RxRenderer(RxRenderer&&) = delete;
-    RxRenderer& operator=(RxRenderer&&) = delete;
+    WindowedRenderer(const WindowedRenderer&) = delete;
+    WindowedRenderer& operator=(const WindowedRenderer&) = delete;
+    WindowedRenderer(WindowedRenderer&&) = delete;
+    WindowedRenderer& operator=(WindowedRenderer&&) = delete;
 
     bool init(int image_width, int image_height, Deltacast::VideoViewer::InputFormat input_format);
     bool start(Deltacast::SharedResources& shared_resources);
@@ -44,7 +47,8 @@ private:
     Deltacast::VideoViewer _monitor;
     std::thread _monitor_thread;
 
-    bool _should_stop = false;
+    std::atomic_bool& _should_stop;
+    std::atomic_bool _monitor_ready;
     std::thread _rendering_loop_thread;
 
     bool monitor(int image_width, int image_height, Deltacast::VideoViewer::InputFormat input_format);